refactor(slack-app): extract rules command handlers into a service
Moves the per-command handlers (`_handle_help`, `_handle_rules_*`,
`_handle_default_repo_*`) out of the Temporal activity into
`services/commands.py` behind a single `dispatch_rules_command` entry
point. The activity now resolves the integration and delegates. Behavior
is unchanged; the picker fallback for `rules add` without a repo stays
inside the workflow because only it can drive the picker signal.
