+++ posthog/temporal/ai/posthog_code_slack_mention.py
@@ -602,165 +602,33 @@ def handle_posthog_code_rules_command_activity(
     user_id: int,
 ) -> PostHogCodeRulesCommandResult:
     from products.slack_app.backend.api import _parse_rules_command
+    from products.slack_app.backend.services.commands import dispatch_rules_command
 
     command = _parse_rules_command(inputs.event.get("text", ""))
     if not command:
         return PostHogCodeRulesCommandResult(status="not_a_command")
+
+    # Picker flow is unique to this workflow; the command service can't drive a
+    # workflow signal, so catch it here before delegating.
+    if command.action == "add" and not command.repository:
+        return PostHogCodeRulesCommandResult(status="needs_picker", pending_rule_text=command.rule_text)
 
     integration = Integration.objects.select_related("team", "team__organization").get(
         id=inputs.integration_id,
         kind="slack-posthog-code",
         integration_id=inputs.slack_team_id,
     )
-    slack = SlackIntegration(integration)
-
-    if command.action == "help":
-        _handle_help(slack, channel, thread_ts)
-    elif command.action == "list":
-        _handle_rules_list(slack, integration, channel, thread_ts)
-    elif command.action == "add":
-        if not command.repository:
-            return PostHogCodeRulesCommandResult(status="needs_picker", pending_rule_text=command.rule_text)
-        _handle_rules_add(slack, integration, channel, thread_ts, user_id, command.rule_text or "", command.repository)
-    elif command.action == "remove":
-        _handle_rules_remove(slack, integration, channel, thread_ts, command.rule_numbers)
-    elif command.action == "deprecated_default_repo":
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text=(
-                "`default repo` commands have been removed. PostHog now selects the repository "
-                "automatically based on your message context — just describe what you need."
-            ),
-        )
-
+    dispatch_rules_command(
+        command,
+        SlackIntegration(integration),
+        integration,
+        channel=channel,
+        thread_ts=thread_ts,
+        slack_user_id=slack_user_id,
+        slack_workspace_id=inputs.slack_team_id,
+        user_id=user_id,
+    )
     return PostHogCodeRulesCommandResult(status="handled")
-
-
-def _handle_help(slack: Any, channel: str, thread_ts: str) -> None:
-    slack.client.chat_postMessage(
-        channel=channel,
-        thread_ts=thread_ts,
-        text=(
-            "*Available commands:*\n\n"
-            "`@PostHog <task description>` — Create a task for the agent to work on\n"
-            "`@PostHog rules list` — Show all routing rules\n"
-            '`@PostHog rules add "description" org/repo` — Add a routing rule\n'
-            '`@PostHog rules add "description"` — Add a routing rule (pick repo from list)\n'
-            "`@PostHog rules remove <number(s)>` — Remove routing rules by number (e.g. `remove 1` or `remove 1,2`)\n"
-            "`@PostHog help` — Show this message\n\n"
-            "You can also reply in an active thread to send follow-up messages to the agent."
-        ),
-    )
-
-
-def _handle_rules_list(slack: Any, integration: Any, channel: str, thread_ts: str) -> None:
-    rules = list(RepoRoutingRule.objects.filter(team_id=integration.team_id).order_by("priority", "id"))
-    if not rules:
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text='No routing rules configured. Add one with `@PostHog rules add "description" [org/repo]`. Omit the repo to pick from a list.',
-        )
-        return
-
-    lines = [f"{i + 1}. {r.rule_text} → `{r.repository}`" for i, r in enumerate(rules)]
-    slack.client.chat_postMessage(
-        channel=channel,
-        thread_ts=thread_ts,
-        text="*Routing rules:*\n" + "\n".join(lines),
-    )
-
-
-def _handle_rules_add(
-    slack: Any,
-    integration: Any,
-    channel: str,
-    thread_ts: str,
-    user_id: int,
-    rule_text: str,
-    repository: str,
-) -> None:
-    from products.slack_app.backend.api import _extract_explicit_repo, _get_full_repo_names
-
-    all_repos = _get_full_repo_names(integration)
-    if not all_repos:
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text="No connected GitHub repositories found for this project.",
-        )
-        return
-
-    matched_repo = _extract_explicit_repo(repository, all_repos)
-    if not matched_repo:
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text=f"Repository `{repository}` is not connected to this project.",
-        )
-        return
-
-    current_max = (
-        RepoRoutingRule.objects.filter(team_id=integration.team_id)
-        .order_by("-priority")
-        .values_list("priority", flat=True)
-        .first()
-    )
-    max_priority = (current_max + 1) if current_max is not None else 0
-    RepoRoutingRule.objects.create(
-        team_id=integration.team_id,
-        rule_text=rule_text,
-        repository=matched_repo,
-        priority=max_priority,
-        created_by_id=user_id,
-    )
-    slack.client.chat_postMessage(
-        channel=channel,
-        thread_ts=thread_ts,
-        text=f"Added rule: {rule_text} → `{matched_repo}`",
-    )
-
-
-def _handle_rules_remove(
-    slack: Any,
-    integration: Any,
-    channel: str,
-    thread_ts: str,
-    rule_numbers: list[int] | None,
-) -> None:
-    if not rule_numbers or any(n < 1 for n in rule_numbers):
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text="Please provide valid rule number(s). Use `@PostHog rules list` to see current rules.",
-        )
-        return
-
-    rules = list(RepoRoutingRule.objects.filter(team_id=integration.team_id).order_by("priority", "id"))
-    invalid = [n for n in rule_numbers if n > len(rules)]
-    if invalid:
-        slack.client.chat_postMessage(
-            channel=channel,
-            thread_ts=thread_ts,
-            text=f"Rule {'number' if len(invalid) == 1 else 'numbers'} {', '.join(f'#{n}' for n in invalid)} {'does' if len(invalid) == 1 else 'do'} not exist. There are {len(rules)} rule(s). Use `@PostHog rules list` to see them.",
-        )
-        return
-
-    # Collect rules to delete (use sorted unique numbers, delete in reverse to keep indices stable)
-    to_delete = sorted(set(rule_numbers), reverse=True)
-    removed: list[str] = []
-    for n in to_delete:
-        rule = rules[n - 1]
-        removed.append(f"#{n}: {rule.rule_text}")
-        rule.delete()
-
-    removed.reverse()
-    slack.client.chat_postMessage(
-        channel=channel,
-        thread_ts=thread_ts,
-        text=f"Removed rule{'s' if len(removed) > 1 else ''} {', '.join(removed)}",
-    )
 
 
 @activity.defn
